Implement node struct
Creates a node struct which will hold a floating point value.

Basic algebraic operations for nodes are supported (+, -, *, and /).

When performing operations using nodes, the result will be stored in it's own node struct which remembers the operation that was used to create it and references to the operands.

Also added some formatting and priting functions for visualization and debug purposes.