Implemented network struct
The old node implementation stored references to the operands (if any) directly inside the node object. This created a variety of issues due to Rusts ownership rules, which made the implementation for calculating the gradients between nodes difficult.

To solve this I introduced a network structure that takes ownership of all the nodes and edges that connect the nodes. This will make ownership more clear, avoid utilizing Rc or RefCell, and hopefully make management of the network much simpler.